Move inference to models pages
